better 7554 ; spi; i2c control; test all outputs selection

+++ src/hardware/Dac7554.h
@@ -32,7 +32,7 @@ public:
      */
     void writeValue(Dac7554Register outputRegister, uint16_t value);
 private:
-    uint8_t _buffer[2]{};
+    uint16_t _buffer[2]{};
 };
